actually correctly properly rename specs-controller to specs-store
