+++ packages/server-ct/src/routes-ct.ts
@@ -5,7 +5,7 @@ import send from 'send'
 import { handle, serve } from '@packages/runner-ct'
 import staticPkg from '@packages/static'
 import ProjectCt from './project-ct'
-import SpecsStore from './specs-controller'
+import SpecsStore from './specs-store'
 
 const debug = Debug('cypress:server:routes')
